Fixes ssue #43. Improve Handling of Edge Cases when Filtering data.
This happened to be a case where the first element returned from
the scan was also in the range we were looking for. Most of the
time, there is some post-scan filtering required since we
sligthly overscan to make up for the scan mechanics and the
data storage mode. As such, this was the first occurence of
such a scenario but it's a good confirmation that I'm ready to
add more unit tests.